+++ admin/assets/manageTravels/update-travels.php
@@ -18,7 +18,7 @@
                 <?php while($donnees = $update_travels_content->fetch()) { ?>
                 <div class="display-pictures-update">
                     <input type="hidden" name="path_pictures" value="../assets/pictures/travels/content/">
-                    <img src="../assets/pictures/travels/content/<?php echo $donnees['picture']; ?>" alt="thumbnail of <?php echo $donnees['title']; ?> - travels article">
+                    <img src="../assets/pictures/travels/assets/<?php echo $donnees['picture']; ?>" alt="thumbnail of <?php echo $donnees['title']; ?> - travels article">
                     <div class="d-flex justify-content-center align-items-center">
                         <label class="btn-file-input-preview">
                             <i class="fa fa-image mr-2"></i><span class="span-no-choice"><?php echo $donnees['picture']; ?></span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="picture">
@@ -27,8 +27,25 @@
                     </div>
                 </div>
                 <div class="other-pictures-update">
-                    <a href="javascript:void(0)" class="add-pictures-link">Modify pictures</a>
-                    <hr>
+                    <!-- <a href="javascript:void(0)" class="add-pictures-link">Modify pictures</a> -->
+                    <div>
+                        <label class="btn-file-input">
+                            <i class="fa fa-image"></i><span class="span-no-choice">Choose picture 1</span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="picture_1">
+                        </label>
+                        <button type="button" class="btn-delete-picture"><i class="far fa-times-circle"></i></button>
+                    </div>
+                    <div>
+                        <label class="btn-file-input">
+                            <i class="fa fa-image"></i><span class="span-no-choice">Choose picture 1</span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="picture_1">
+                        </label>
+                        <button type="button" class="btn-delete-picture"><i class="far fa-times-circle"></i></button>
+                    </div>
+                    <div>
+                        <label class="btn-file-input">
+                            <i class="fa fa-image"></i><span class="span-no-choice">Choose picture 1</span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="picture_1">
+                        </label>
+                        <button type="button" class="btn-delete-picture"><i class="far fa-times-circle"></i></button>
+                    </div>
                     <div>
                         <label class="btn-file-input">
                             <i class="fa fa-image"></i><span class="span-no-choice">Choose picture 1</span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="picture_1">